requirements.txt
# Discord
discord.py>=2.3,<3.0

# Web
fastapi>=0.110,<1.0
uvicorn[standard]>=0.29,<1.0
pydantic>=2.0,<3.0
pydantic-settings>=2.0,<3.0

# Database
aiosqlite>=0.20,<1.0

# HTTP (for email/API calls later)
httpx>=0.27,<1.0

# Testing
pytest>=8.0
pytest-asyncio>=0.23
